moved go files to separate directory
