+++ app/backup.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"crypto/sha256"
@@ -136,8 +136,8 @@ func loadExisting(overridePath string, gridDir string, game *Game, artStyleExten
 
 				// remove overlay
 				game.OverlayImageBytes = nil
+			}
 		}
-	}
 
-}
+	}
 }
